requirements.txt
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
dnspython==1.16.0
gunicorn==20.0.4
numpy==1.17.4
pandas==0.25.3
psycopg2==2.8.4
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1